[third-party/py-spy/0.3.14] Only check main thread stack trace in address checker
Summary:
The address checker, used to locate the interpreter in the binary,
does a final check by grabbing the stack traces for all threads from
the interpreter address, to make sure the address is sane.

However, this check isn't gated by the process lock that is normally
used to get stack traces, and so is prone to race conditions.  As this
function doesn't really need to look for all threads, just have it get
the stack trace for the main thread, so that it's significantly less likely
to fail.

Test Plan: Ran with a thread stress testing.

Reviewers: bmaurer

Reviewed By: bmaurer

Subscribers: mzlee

Differential Revision: https://phabricator.intern.facebook.com/D50869121